Revert "Merge remote-tracking branch 'origin/main' into devel-main"
This reverts commit 890b59a30c414c59cfa74776f92fe138df74f13f, reversing
changes made to f5da356bb92421d3b5964c2af1cab87c8a970030.

This undoes the merging of the freebsd 25 upstream, therefore restoring the source code of the freebsd kernel used in pfSense 2.8.0

+++ sys/fs/smbfs/smbfs_vnops.c
@@ -1052,7 +1052,7 @@ smbfs_lookup(struct vop_lookup_args *ap)
 	struct smbfattr fattr, *fap;
 	struct smb_cred *scred;
 	char *name = cnp->cn_nameptr;
-	uint64_t flags = cnp->cn_flags;
+	int flags = cnp->cn_flags;
 	int nameiop = cnp->cn_nameiop;
 	int nmlen = cnp->cn_namelen;
 	int error, islastcn, isdot;
